rollback Future instrumentation to Kamon 1.x behavior
