+++ instrumentation/kamon-akka-http/src/main/resources/reference.conf
@@ -238,6 +238,8 @@ kanela.modules {
     name = "Akka HTTP Instrumentation"
     description = "Provides context propagation, distributed tracing and HTTP client and server metrics for Akka HTTP"
 
+    # Add "kamon.instrumentation.akka.http.FastFutureInstrumentation" to the instrumentation list if you need to
+    # use the Future Chaining instrumentation in Kamon 2.2.x. It will be fully removed in Kamon 2.3.0.
     instrumentations = [
       "kamon.instrumentation.akka.http.AkkaHttpServerInstrumentation"
       "kamon.instrumentation.akka.http.AkkaHttpClientInstrumentation"
