+++ instrumentation/kamon-akka-http/src/main/scala-2.11/kamon/instrumentation/akka/http/AkkaHttpServerInstrumentation.scala
@@ -78,6 +78,9 @@ class AkkaHttpServerInstrumentation extends InstrumentationBuilder {
     .intercept(method("redirect"), classOf[ResolveOperationNameOnRouteInterceptor])
     .intercept(method("failWith"), classOf[ResolveOperationNameOnRouteInterceptor])
 
+}
+
+class FastFutureInstrumentation extends InstrumentationBuilder {
 
   /**
     * This allows us to keep the right Context when Futures go through Akka HTTP's FastFuture and transformantions made
@@ -304,7 +307,9 @@ object FastFutureTransformWithAdvice {
       catch { case NonFatal(e) => FastFuture.failed(e) }
 
     // If we get a FulfilledFuture or ErrorFuture, those will have the HasContext mixin,
-    // otherwise we are getting a regular Future which has the context mixed into its value.
+    // otherwise we are getting a regular Future that might have a context mixed into its
+    // value if the Future Chaining instrumentation is manually enabled (it was deprecated
+    // in Kamon 2.1.21).
     if(future.isInstanceOf[HasContext])
       zuper.call()
     else {
@@ -316,18 +321,25 @@ object FastFutureTransformWithAdvice {
             case Failure(e) => p completeWith strictTransform(e, f)
           }(ec)
           p.future
-        case Some(value) =>
-          // This is possible because of the Future's instrumentation
-          val futureContext = value.asInstanceOf[HasContext].context
-          val scope = Kamon.storeContext(futureContext)
 
-          val transformedFuture = value match {
+        case Some(value) =>
+          value match {
+            case hc: HasContext =>
+              // The Future's value (Try[A]) might have a Context stored if the
+              // Future Chaining instrumentation is enabled.
+              val scope = Kamon.storeContext(hc.context)
+
+              val transformedFuture = value match {
+                case Success(a) => strictTransform(a, s)
+                case Failure(e) => strictTransform(e, f)
+              }
+
+              scope.close()
+              transformedFuture
+
             case Success(a) => strictTransform(a, s)
             case Failure(e) => strictTransform(e, f)
           }
-
-          scope.close()
-          transformedFuture
       }
     }
   }
