+++ instrumentation/kamon-akka-http/src/main/scala-2.13/kamon/instrumentation/akka/http/AkkaHttpServerInstrumentation.scala
@@ -80,6 +80,16 @@ class AkkaHttpServerInstrumentation extends InstrumentationBuilder {
 
 
   /**
+    * Akka-http 10.1.x compatibility.
+    */
+
+  onType("akka.http.scaladsl.Http2Ext")
+    .advise(method("bindAndHandleAsync") and isPublic(), classOf[Http2ExtBindAndHandleAdvice])
+}
+
+class FastFutureInstrumentation extends InstrumentationBuilder {
+
+  /**
     * This allows us to keep the right Context when Futures go through Akka HTTP's FastFuture and transformantions made
     * to them. Without this, it might happen that when a Future is already completed and used on any of the Futures
     * directives we might get a Context mess up.
@@ -92,13 +102,6 @@ class AkkaHttpServerInstrumentation extends InstrumentationBuilder {
 
   onType("akka.http.scaladsl.util.FastFuture$")
     .intercept(method("transformWith$extension").and(takesArguments(4)), FastFutureTransformWithAdvice)
-
-  /**
-    * Akka-http 10.1.x compatibility.
-    */
-
-  onType("akka.http.scaladsl.Http2Ext")
-    .advise(method("bindAndHandleAsync") and isPublic(), classOf[Http2ExtBindAndHandleAdvice])
 }
 
 trait HasMatchingContext {
@@ -328,18 +331,25 @@ object FastFutureTransformWithAdvice {
             case Failure(e) => p completeWith strictTransform(e, f)
           }(ec)
           p.future
-        case Some(value) =>
-          // This is possible because of the Future's instrumentation
-          val futureContext = value.asInstanceOf[HasContext].context
-          val scope = Kamon.storeContext(futureContext)
 
-          val transformedFuture = value match {
+        case Some(value) =>
+          value match {
+            case hc: HasContext =>
+              // The Future's value (Try[A]) might have a Context stored if the
+              // Future Chaining instrumentation is enabled.
+              val scope = Kamon.storeContext(hc.context)
+
+              val transformedFuture = value match {
+                case Success(a) => strictTransform(a, s)
+                case Failure(e) => strictTransform(e, f)
+              }
+
+              scope.close()
+              transformedFuture
+
             case Success(a) => strictTransform(a, s)
             case Failure(e) => strictTransform(e, f)
           }
-
-          scope.close()
-          transformedFuture
       }
     }
   }
