+++ instrumentation/kamon-executors/src/main/resources/reference.conf
@@ -27,7 +27,6 @@ kanela.modules {
       "^com.sun.tools.*",
       "^sbt.internal.*",
       "^com.intellij.rt.*",
-      "^scala.concurrent.*",
       "^org.jboss.netty.*",
       "^com.google.common.base.internal.Finalizer",
       "^kamon.module.*",
