+++ instrumentation/kamon-scala-future/src/main/resources/reference.conf
@@ -21,10 +21,16 @@ kamon.instrumentation.futures.scala {
 }
 
 kanela.modules {
+  executor-service {
+    within += "scala.concurrent.impl.CallbackRunnable"
+    within += "scala.concurrent.impl.Future\\$PromiseCompletingRunnable"
+    within += "scala.concurrent.impl.Promise\\$Transformation"
+  }
 
   scala-future {
-    name = "Scala Future Instrumentation"
+    name = "Scala Future Chaining Instrumentation (Deprecated)"
     order = 1
+    enabled = false
     description = "Deprecated since Kamon 2.1.21, will be removed in Kamon 2.3.0. Provides automatic context propagation to the thread executing a Scala Future's body and callbacks"
     instrumentations = [
       "kamon.instrumentation.futures.scala.FutureChainingInstrumentation"
